requirements.txt
certifi==2019.6.16
chardet==3.0.4
facenet-pytorch==0.1.0
idna==2.8
numpy==1.17.1
opencv-python==4.1.1.26
pandas==0.25.1
Pillow==6.1.0
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
scipy==1.3.1
six==1.12.0
torch==1.2.0
torchvision==0.4.0
urllib3==1.25.3
